post content fetch and update

+++ leetcomp/services.py
@@ -1,4 +1,4 @@
-from contextlib import contextmanager
+from importlib.metadata import requires
 import json
 import math
 import os
@@ -8,23 +8,17 @@ import time
 from typing import Any, Dict, List, Set, Tuple, Union
 
 from loguru import logger
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
 from tqdm import trange
 
-from .models import Base, Posts
-from .queries import COMP_POSTS_DATA_QUERY
-from .utils import get_today
+from .models import Posts
+from .queries import COMP_POSTS_DATA_QUERY, COMP_POST_CONTENT_DATA_QUERY
+from .utils import get_today, session_scope
+
 
 CACHE_DIR = ".cache"
 DATA_TYPE = Tuple[List[Dict[str, Union[int, str]]], int, bool]
 LEETCODE_GRAPHQL_URL = "https://leetcode.com/graphql"
 
-# sqlite engine
-engine = create_engine("sqlite:///posts.db")
-Session = sessionmaker()
-Session.configure(bind=engine)
-Base.metadata.create_all(engine)
 
 # cache dir; removes the data cached before today's date
 if not os.path.exists(CACHE_DIR):
@@ -32,27 +26,14 @@ if not os.path.exists(CACHE_DIR):
 else:
     today = get_today()
     for f in os.listdir(CACHE_DIR):
-        if not f.startswith(today):
+        if not f.startswith(today) and not f.startswith("post_id"):
             logger.info(f"Removing old cache {f}")
             os.remove(f"{CACHE_DIR}/{f}")
 
 
-@contextmanager
-def session_scope() -> sessionmaker:
-    session = Session()
-    try:
-        yield session
-        session.commit()
-    except:
-        session.rollback()
-        raise
-    finally:
-        session.close()
-
-
 def _get_post_ids_in_db() -> Set[str]:
     with session_scope() as session:
-        return set([r[0] for r in session.query(Posts.id).all()])
+        return set([r.id for r in session.query(Posts.id).all()])
 
 
 def _validate_comp_posts_response(response: requests.Response) -> None:
@@ -64,32 +45,62 @@ def _validate_comp_posts_response(response: requests.Response) -> None:
     ), f"`categoryTopicList` not in {response_json['data']}"
 
 
-def _get_all_comp_posts(query: str, posts_cache_dir: str) -> Tuple[Dict[str, Any], bool]:
-    if os.path.exists(posts_cache_dir):
-        with open(posts_cache_dir, "r") as f:
+def _validate_post_content_response(response: requests.Response) -> None:
+    assert response.status_code == 200
+    response_json = response.json()
+    assert "data" in response_json, f"`data` not in {response_json}"
+    assert "topic" in response_json["data"], f"`topic` not in {response_json['data']}"
+    assert (
+        "post" in response_json["data"]["topic"]
+    ), f"`post` not in {response_json['data']['topic']}"
+    assert (
+        "content" in response_json["data"]["topic"]["post"]
+    ), f"`post` not in {response_json['data']['topic']['post']}"
+
+
+def _get_all_comp_posts(query: str, posts_cache_path: str) -> Tuple[Dict[str, Any], bool]:
+    if os.path.exists(posts_cache_path):
+        with open(posts_cache_path, "r") as f:
             posts_data = json.load(f)
-        used_cache = True
+        cache_is_used = True
     else:
         response = requests.post(LEETCODE_GRAPHQL_URL, json=query)
         _validate_comp_posts_response(response)
-        with open(posts_cache_dir, "w") as f:
+        with open(posts_cache_path, "w") as f:
             json.dump(response.json(), f)
         posts_data = response.json()
-        used_cache = False
+        cache_is_used = False
 
-    return posts_data, used_cache
+    return posts_data, cache_is_used
 
 
-def _get_filterted_posts(skip: int = 0, first: int = 15) -> DATA_TYPE:
+def _get_content_from_post_id(query: str, post_content_cache_path: str) -> str:
+    if os.path.exists(post_content_cache_path):
+        with open(post_content_cache_path, "r") as f:
+            post_content = json.load(f)
+        cache_is_used = True
+    else:
+        response = requests.post(LEETCODE_GRAPHQL_URL, json=query)
+        _validate_post_content_response(response)
+        with open(post_content_cache_path, "w") as f:
+            json.dump(response.json(), f)
+        post_content = response.json()
+        cache_is_used = False
+
+    return post_content, cache_is_used
+
+
+def _get_info_from_posts(skip: int = 0, first: int = 15) -> DATA_TYPE:
     COMP_POSTS_DATA_QUERY["variables"]["skip"] = skip
     COMP_POSTS_DATA_QUERY["variables"]["first"] = first
-    posts_cache_dir = f"{CACHE_DIR}/{get_today()}_posts_data_{skip}_{first}.json"
+    posts_cache_path = f"{CACHE_DIR}/{get_today()}_posts_data_{skip}_{first}.json"
 
-    data, used_cache = _get_all_comp_posts(COMP_POSTS_DATA_QUERY, posts_cache_dir)
+    data, cache_is_used = _get_all_comp_posts(COMP_POSTS_DATA_QUERY, posts_cache_path)
     data = data["data"]["categoryTopicList"]
     if not data or "edges" not in data:
         logger.warning(f"Missing data for skip {skip}, first {first}")
         return [], 0, False
+
     filtered_data = [
         {
             "id": d["node"]["id"],
@@ -102,44 +113,86 @@ def _get_filterted_posts(skip: int = 0, first: int = 15) -> DATA_TYPE:
         for d in data["edges"]
     ]
 
-    return filtered_data, data["totalNum"], used_cache
+    return filtered_data, data["totalNum"], cache_is_used
 
 
-def _get_new_posts(data: DATA_TYPE, old_posts: Set[str]) -> List[Dict[str, Union[int, str]]]:
+def _get_content_from_post(post_id: str) -> str:
+    COMP_POST_CONTENT_DATA_QUERY["variables"]["topicId"] = int(post_id)
+    post_content_cache_path = f"{CACHE_DIR}/post_id_{post_id}.json"
+
+    data, cache_is_used = _get_content_from_post_id(
+        COMP_POST_CONTENT_DATA_QUERY, post_content_cache_path
+    )
+    data = data["data"]["topic"]["post"]["content"]
+    if not data:
+        logger.warning(f"Missing content for post_id {post_id}")
+        return "", False
+
+    return data, cache_is_used
+
+
+def _get_new_posts(
+    data: DATA_TYPE, old_posts: Set[str]
+) -> List[Dict[str, Union[int, str]]]:
     return [d for d in data if d["id"] not in old_posts]
 
 
-def _update_old_posts(old_posts: Set[str], new_data: DATA_TYPE):
+def _update_old_post_ids(old_posts: Set[str], new_data: DATA_TYPE):
     for d in new_data:
         old_posts.add(d["id"])
 
 
-def get_and_save_all_comp_posts() -> None:
-    stride_size = 15
+def _get_post_ids_without_content() -> List[str]:
+    with session_scope() as session:
+        return [r.id for r in session.query(Posts.id).filter(Posts.content == "").all()]
+
+
+def get_posts_meta_info() -> None:
+    n_posts_per_req = 15
     start = 0
-    old_posts = _get_post_ids_in_db()
+    old_post_ids = _get_post_ids_in_db()
     try:
         # fetching the first page separately to get totalNum pages
-        data, n_posts, used_cache = _get_filterted_posts(start, stride_size)
-        new_data = _get_new_posts(data, old_posts)
+        data, n_posts, cache_is_used = _get_info_from_posts(start, n_posts_per_req)
+        new_data = _get_new_posts(data, old_post_ids)
         with session_scope() as session:
             session.add_all([Posts(**d) for d in new_data])
-            _update_old_posts(old_posts, new_data)
-        n_pages = math.ceil(n_posts / stride_size)
-        logger.info(f" Found {n_posts} posts({n_pages} pages) ".center(34, "*"))
+            _update_old_post_ids(old_post_ids, new_data)
+        n_pages = math.ceil(n_posts / n_posts_per_req)
+        logger.info(f"Found {n_posts} posts({n_pages} pages)")
         # fetching the rest of the pages
         with trange(1, n_pages + 1) as t:
             for page_no in t:
-                sleep_for = 0 if used_cache else random.random() + 0.5
+                sleep_for = 0 if cache_is_used else random.random() + 0.5
                 time.sleep(sleep_for)
-                start += stride_size
-                data, _, used_cache = _get_filterted_posts(start, stride_size)
-                new_data = _get_new_posts(data, old_posts)
+                start += n_posts_per_req
+                data, _, cache_is_used = _get_info_from_posts(start, n_posts_per_req)
+                new_data = _get_new_posts(data, old_post_ids)
                 with session_scope() as session:
                     session.add_all([Posts(**d) for d in new_data])
-                    _update_old_posts(old_posts, new_data)
+                    _update_old_post_ids(old_post_ids, new_data)
                 t.set_description(f"Page {page_no}")
                 t.set_postfix(slept_for=sleep_for)
+    except KeyboardInterrupt:
+        session.commit()
+        session.close()
+
+
+def update_posts_content_info() -> None:
+    post_ids_without_content = _get_post_ids_without_content()
+    try:
+        cache_is_used = False
+        with trange(len(post_ids_without_content)) as t:
+            for ix in t:
+                post_id = post_ids_without_content[ix]
+                sleep_for = 0 if cache_is_used else random.random() + 0.5
+                time.sleep(sleep_for)
+                post_content, cache_is_used = _get_content_from_post(post_id)
+                with session_scope() as session:
+                    post = session.query(Posts).filter(Posts.id == post_id).first()
+                    post.content = post_content
+                t.set_description(f"PostId {post_id}")
+                t.set_postfix(sleep_for=sleep_for)
     except KeyboardInterrupt:
         session.commit()
         session.close()
